requirements.txt
Flask==3.0.0
pandas==2.1.4
transformers==4.36.2
torch==2.1.2
google-generativeai==0.3.2
Werkzeug==3.0.1
Jinja2==3.1.2
numpy==1.24.4
scikit-learn==1.3.2
requests==2.31.0
